Fixed a compile problem that occurs on certain older 32-bit Ubuntu releases.

+++ clientlib/cryptcurve25519.c
@@ -693,7 +693,7 @@ _cryptcurve25519_updatedata(Frame* f,			///< Frame to marshall
 	// Encrypt in-place [we previously allocated enough space for authentication info]
 	crypto_box_easy(tlvval+cyphertextoffset, tlvval+plaintextoffset, plaintextsize
 	,	nonce, self->public_key->public_key, self->private_key->private_key);
-	DEBUGMSG4("cypher offset versus tlvstart: %ld", tlvval+cyphertextoffset-(guint8*)tlvstart);
+	DEBUGMSG4("cypher offset versus tlvstart: %ld", (long)(tlvval+cyphertextoffset-(guint8*)tlvstart));
 	DEBUGCKSUM4("cypher text checksum:", tlvval+cyphertextoffset, plaintextsize+crypto_box_MACBYTES);
 	set_generic_tlv_type(tlvstart, self->baseclass.baseclass.type, pktend);
 	set_generic_tlv_len(tlvstart, tlvsize, pktend);
